+++ config/initializers/assets.rb
@@ -11,4 +11,11 @@ Rails.application.config.assets.version = "1.0"
 # folder are already added.
 # Rails.application.config.assets.precompile += %w( admin.js admin.css )
 Rails.application.config.assets.precompile += %w(bootstrap.min.js popper.js)
+Rails.application.config.assets.precompile << "chart.js/dist/chart.umd.js"
+
+# Precompile all Popper.js ESM files
+node_modules_path = Rails.root.join("node_modules")
+Rails.application.config.assets.precompile += Dir[node_modules_path.join("@popperjs/core/dist/esm/**/*.js")].map do |f|
+  f.sub("#{node_modules_path}/", "")
+end
 Rails.application.config.assets.paths << Rails.root.join("node_modules")